+++ src/unicorn.js
@@ -1,0 +1,13 @@
+export const unicornConfig = {
+  name: 'unicorn',
+  rules: {
+    'unicorn/catch-error-name': [
+      'error',
+    ],
+    'unicorn/consistent-destructuring': 'error',
+    'unicorn/consistent-empty-array-spread': 'error',
+    'unicorn/consistent-existence-index-check': 'error',
+    'unicorn/error-message': 'error',
+    'unicorn/expiring-todo-comments': 'error',
+  },
+}